refactor(downloader): 独立运行改造
- 移除 CodeFirst 自动建表，改为 EnsureTablesCreated 首次运行自动创建
- 启动时连接 DFApp 后端失败不再阻止程序启动
- 添加 /api/connection/reconnect 端点用于重新连接
- 修正登录 API 路由为 /api/app/account/login
- 修正登录响应解析匹配服务器格式
- 实体 Id 类型从 long 改为 int（SQLite AUTOINCREMENT 限制）
- SqlSugar 改为 SqlSugarCore
- 添加 SPA fallback 路由支持
- 更新 README 简化快速开始步骤

+++ DFApp.Downloader/src/DFApp.Downloader.Core/Engine/DownloadEngine.cs
@@ -98,17 +98,17 @@ public class DownloadEngine
     private readonly DownloaderSettings _settings;
     private readonly HttpClient _httpClient;
     private readonly ILogger<DownloadEngine> _logger;
-    private readonly ConcurrentDictionary<long, CancellationTokenSource> _activeDownloads = new();
+    private readonly ConcurrentDictionary<int, CancellationTokenSource> _activeDownloads = new();
     private readonly SemaphoreSlim _concurrencySemaphore;
 
     /// <summary>下载进度事件</summary>
     public event Action<DownloadProgress>? OnProgress;
 
     /// <summary>下载完成事件</summary>
-    public event Action<long>? OnDownloadCompleted;
+    public event Action<int>? OnDownloadCompleted;
 
     /// <summary>下载失败事件</summary>
-    public event Action<long, string>? OnDownloadFailed;
+    public event Action<int, string>? OnDownloadFailed;
 
     public DownloadEngine(DownloaderSettings settings, HttpClient httpClient, ILogger<DownloadEngine> logger)
     {
@@ -150,7 +150,7 @@ public class DownloadEngine
     /// <summary>
     /// 暂停下载
     /// </summary>
-    public void PauseDownload(long itemId)
+    public void PauseDownload(int itemId)
     {
         if (_activeDownloads.TryGetValue(itemId, out var cts))
         {
